Create Intake Subsystem.
Currently assumes a 4 motor design, however this is not final.
There is no implementation for the cube/cone in/out methods as the final motor/chain layout of the intake is unclear.
Uses current limit for neo 550s as it is likely we will use them over the regular neos.
getState will require a yet unknown configuration of break beam switches so it is not yet implemented.
Remove rouge patch file.