Better overextension pervention
Instead of telling the wrist to tuck in and stuff when we are at risk of overextending (worked, but the wrist was too sluggish) we instead check if moving the mechanism (wrist, arm rot, extension) in a certain direction will cause an overextension (for the wrist this is if moving it in the direction will potetialy pass our saftey line, while for the arm it checks if it is past the safteyline (a little bit under the extension limit) and if moving it in that direction would make it worse.

Note, right now this is only enabled for manual control, as it will be painful to implement for the presets (they shouldn't overextend anyways) also us setting the arm to 0 when moving it up/down would cause an overextension might cause an issue as the spring/gravity might pull the arm in that direction, we could probably have setRotationSpeed return weather or not that direction would cause overextension and cancel manual control so the default command will hold the arm in place, but I highly doubt this is necessary as the inspector only cared about the wrist and extension causing an overextension, and not the arm rot, but we can implement it if our tulsa inspector is more strict.